+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file, in reverse 
 - The fps counter was refactored ([#69](https://github.com/tuupola/hagl/pull/69)).
 - The aps counter was refactored ([#69](https://github.com/tuupola/hagl/pull/70)).
 
+### Added
+- XYX and XYW variants for hline ([#82](https://github.com/tuupola/hagl/pull/82)).
+
 ## 0.5.0 - 2022-05-19
 
 Initial release.